Some zstr-related changes
