+++ LICENSE
@@ -29,10 +29,10 @@ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 Additionally this project distributes 3 libraries with their own respective
-licenses which can be found in their code and attached license files. 
-These libraries and their licenses are as follows:
+licenses which can be found in their code and attached license files. These
+libraries and their licenses are as follows:
 
-* Picojson (fem/picojson.h) - Custom 2 clause license
-* Catch++  (tests/unit/catch.hpp, tests/unit/LICENSE_1_0.txt) - Boost 1.0 license
-* Zstr (general/zstr.hpp, general/zstr_strict_fstream.hpp) - MIT license
+* Picojson (fem/picojson.h) -- Custom 2-clause license
+* Catch++  (tests/unit/catch.hpp) -- Boost 1.0 license
+* Zstr (general/zstr.hpp) -- MIT license
