+++ mfem.hpp
@@ -28,6 +28,7 @@
 #include "general/osockstream.hpp"
 #include "general/socketstream.hpp"
 #include "general/optparser.hpp"
+#include "general/zstr.hpp"
 #include "general/version.hpp"
 #include "general/globals.hpp"
 #ifdef MFEM_USE_MPI
